\name{ar1mat}
\alias{ar1mat}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ create correlation matrix with AR(1) structure }
\description{
create correlation matrix with AR(1) structure
}
\usage{
ar1mat(n=4, rho=0.5)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ dimension of square matrix }
  \item{rho}{ correlation parameter value }
}
\details{
returns patterned matrix; see also nlme
}
\value{
patterned matrix
}

%\examples{
%}
\keyword{ models }% at least one, from doc/KEYWORDS


