\name{csmat}
\alias{csmat}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ create correlation matrix of exchangeable structure }
\description{
create correlation matrix of exchangeable (compound symmetric) structure
}
\usage{
csmat(n=4, rho=0.5)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ dimension of square matrix }
  \item{rho}{ correlation parameter }
}
%\details{
%}
%\value{
%}
%\references{  }



\examples{
csmat(n=4, rho=.5)
}
\keyword{ models }


