\name{DEMOdf}
\alias{DEMOdf}
\non_function{}
\title{ simple demo data frame }
\usage{data(DEMOdf)}
\description{
The \code{DEMOdf} data frame has 40 rows and 7 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{YO}{a numeric vector}
    \item{x2}{a numeric vector}
    \item{x3}{a numeric vector}
    \item{x4}{a numeric vector}
    \item{x5}{a numeric vector}
    \item{ID}{a numeric vector}
    \item{TIM}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
data(DEMOdf)
}
\keyword{datasets}

