\name{print.yagsResult}
\alias{print.yagsResult}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ write a report of a yags GEE solution on standard output }
\description{
invokes various component functions to report on a GEE solution
obtained by yags
}
\usage{
print.yagsResult(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ object of class yagsResult }
  \item{\dots}{ other parameters, required for generic show method }
}
\details{
assembles results and prints
}
\value{
}
\author{ VJ Carey }
\note{ invoked via the generic show method for class yagsResult }

\examples{
}
\keyword{ models }% at least one, from doc/KEYWORDS

