\name{WTS}
\alias{WTS}
\description{ a vector of weights for reproducible testing }
\non_function{}
\title{ a vector of weights for use with DEMOdf }
\usage{data(WTS)}
\format{
  The format is:
 num [1:40] 1.051 0.819 1.213 0.941 1.252 ...
}
%\source{
%}
%\references{
%}
\examples{
data(WTS)
## maybe str(WTS) ; plot(WTS) ...
}
\keyword{datasets}

