\name{yags.wcorReport}
\alias{yags.wcorReport}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ report on working correlation elements of a GEE solution; not
intended to be called directly. }
\description{ report on working correlation elements of a GEE solution; not
intended to be called directly. }
\usage{
yags.wcorReport(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class yagsResult }
  \item{\dots}{ not currently used }
}
%\details{
%}
\value{
A string including parameter estimates and standard errors
where available.
}
%\references{ }
\author{ VJ Carey, stvjc@gauss.med.harvard.edu }
%\note{ }

%\seealso{ }

%\examples{ }
\keyword{ models }
%\keyword{ longitudinal data }% at least one, from doc/KEYWORDS

